pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "go2-terrain-curriculum"
version = "0.1.0"
description = "A benchmark-oriented quadruped locomotion starter focused on terrain curriculum and reward ablations."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "yijian-stack" }]
dependencies = [
  "numpy>=1.26,<3.0",
  "PyYAML>=6.0,<7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.6,<1.0"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

